Add com.netflix.hollow.core.type support for Byte
Rather than generating Byte classes when we encouunter a
Byte as part of another Type, introduce a top-level Byte
type as part of the core hollow types. Bytes are stored
as integers under the hood, and the generated classes will
use integers, similar to how a primitive byte gets converted
to an int.
Adding support for byte to the hollow schema is out of scope
for this PR.