Fix for #166 - show `[null]` for null nodes in diff/history UI
For cases where we have a node that isn't a leaf (i.e. it
represents an object), call out the fact that it is null,
instead of ignoring that in the history and diff UIs. This
is important for seeing the difference between a node that
is null and a node that has no children (e.g. an empty list,
set, or map).