Allow creating HollowConsumers for testing purposes
Add TestHollowConsumer, TestBlobRetriever, and
TestAnnouncementWatcher to facilitate creating
HollowConsumer objects for use in unit tests.
Generally speaking, a TestHollowConsumer is constructed
with a TestAnnouncementWatcher and TestBlobRetriever, and
can be manipulated load a snapshot as represented by a
HollowWriteStateEngine. Support for loading deltas will come
later.
As part of this effort, refactor
HollowReadStateEngineBuilder to extract a
HollowWriteStateEngineBuilder.