Reflect actual performance of low cardinality indexes
I realised that small indexes are going to just sit in cache in a tight loop and the actual performance would be misrepresented. This adds a naieve target size and entry overhead to preallocate indexes which are selected at random, to make sure we're going to memory for each iteration.