Specify runCycle to return the latests announced version.
Previously if population resulted in no changes a version
was returned that would be the announced version if there had
been changes.
The change in behaviour is consistent with if the producer is not
the primiary producer
The version may be consistently checked against consumers and their
current version.  The version reported to event listeners has not
changed.