Disable the delta update of a primary key index.
Delta updates continue to be problematic with reports
of matches unexpectedly returning no result.  This suggests
an index may be in a currupt but undetected state (until, say,
a delta update is applied).  For now it is consideted best
to disable such updates while attempting to investigate the
issue further (a system property has been defined to enable
delta updates for such purposes).  As a consequence this may
increase CPU time (but not memory consumption).