Maintain correct read states after integrity check (#429)
Previously, we were swapping the current and pending read states
(after applying the delta and reverse delta, respectively),
for efficiency. However, this is not the correct
thing to do when the two read states don't have the same
schemas. For those cases, apply the delta back onto the
pending state, and apply the reverse delta back onto the
current state.