add API contract tagging annotations
Annotations to use on classes and/or methods to indicate our intended
contract for API stability.

Ideally all types with the `public` access modifier will be annotated.
It may also useful to annotate package-private classes with
`@Internal`; this clearly signals intent and eliminates the ambiguity
of whether we "forgot" to make it public.

Hollow supports Java 8 for the forseable future; so we cannot use
the Java module system. Additionally Java modules don't offer the
distinction that `@PublicApi` versus `@PublicSpi` do.