Deal with race condition between snapshot application and HollowConsumer#getAPI
If a user calls HollowConsumer#getAPI in the middle of a
snapshot update, they will get back null if the hollowDataHolder
volatile field is assigned but the snapshot is not yet applied.
To work around this, only set the volatile field after the
new API is initialized (and so the snapshot is processed)
but before refresh handlers are called.