Update PrimaryKeyIndex to use HollowDataAccess instead of HollowObjectTypeReadState. This allows PrimaryKeyIndex to opt-in to Object Longevity by using the HollowDataAccess from a Consumer rather than the HollowObjectTypeReadState.
HollowObjectTypeReadState is mutable and gets updated with each delta.  If a PrimaryKeyIndex does NOT listen to updates, then it will become stale over time. The internal index structure will remain unchanged, but the HollowPrimaryKeyValueDeriver will start returning incorrect keys as the ordinals get re-used and re-assigned to new records.

By migrating to HollowDataAccess, users can choose to use the HollowObjectTypeReadState (which stays up to date) or use HollowApi.getDataAccess() which may be object-longevity-aware.