Update HollowHashIndex to use HollowDataAccess instead of HollowObjectTypeReadState. This allows HollowHashIndex to opt-in to Object Longevity by using the HollowDataAccess from a Consumer rather than the HollowObjectTypeReadState.
HollowObjectTypeReadState is mutable and gets updated with each delta.  If a HollowHashIndex does NOT listen to updates, then it will become stale over time. Unfortunately, when using the StateEngine, the staleness causes the ordinals to mismatch against the state the index was built against.

By migrating to HollowDataAccess, users can choose to use the HollowObjectTypeReadState (which stays up to date) or use HollowApi.getDataAccess() which may be object-longevity-aware.