Add a new HollowUniqueKeyIndex that supports object longevity. Update HollowHashIndex to support object longevity. (#567)
* Update PrimaryKeyIndex to use HollowDataAccess instead of HollowObjectTypeReadState. This allows PrimaryKeyIndex to opt-in to Object Longevity by using the HollowDataAccess from a Consumer rather than the HollowObjectTypeReadState.

HollowObjectTypeReadState is mutable and gets updated with each delta.  If a PrimaryKeyIndex does NOT listen to updates, then it will become stale over time. The internal index structure will remain unchanged, but the HollowPrimaryKeyValueDeriver will start returning incorrect keys as the ordinals get re-used and re-assigned to new records.

By migrating to HollowDataAccess, users can choose to use the HollowObjectTypeReadState (which stays up to date) or use HollowApi.getDataAccess() which may be object-longevity-aware.

* Update HollowHashIndex to use HollowDataAccess instead of HollowObjectTypeReadState. This allows HollowHashIndex to opt-in to Object Longevity by using the HollowDataAccess from a Consumer rather than the HollowObjectTypeReadState.

HollowObjectTypeReadState is mutable and gets updated with each delta.  If a HollowHashIndex does NOT listen to updates, then it will become stale over time. Unfortunately, when using the StateEngine, the staleness causes the ordinals to mismatch against the state the index was built against.

By migrating to HollowDataAccess, users can choose to use the HollowObjectTypeReadState (which stays up to date) or use HollowApi.getDataAccess() which may be object-longevity-aware.

* Fix type data accessor lookups for collection and map.

* Add addition tests to HollowHashIndexTest.

* Refactor out most of the common code between HollowPrimaryKeyIndex and HollowUniqueKeyIndex. Add test for longevity on hash index. Merge HollowPrimaryKeyValueDeriver2.java into HollowUniqueKeyIndex.

* Take out refactoring of HollowPrimaryKeyIndex since the long term plan is to deprecate it.

* Fix checkstyle.

* Fix checkstyle.

* Address PR comments.

Add deprecation to internal interface.

Suppress a few warnings.

Restore API to the original signature.