handle null pkeys in explorer ui
Null primary keys are not supported, however when a null primary key is
present, browsing a selected type in the Hollow Explorer UI will fail
with an OOB exception.

Handle null primary keys so that they are still displayed in the
selected type page, with "null" substituted in-place of missing fields.